requirements.txt
pandas==1.0.3
ics==0.7
python-slugify==4.0.0
